Replaced the old codebase in its entirety
This commit replaces the previous btree library, which was a
fork of https://github.com/google/btree.

The old code can be found in the 'google-fork' branch and/or the
v0.1.0 tag. The old commits are still available to keep in sync
with the Go module system or other dependency managers.