Preparing for Typed Parameters
This commit is a major rewrite, but has no new features or
bugfixes. It's purpose is to begin preparing the codebase for
Go 1.18, which will include generics.

The new "internal" directory contains the newest implementation
of my B-tree with hardcoded parameters at the top of the file.

There are a couple changes to the implentation. The big one is
the Get, Set, Delete functions include two values. The first is
the item value and the second is a boolean which indicates that
the value is not empty. This is similar to using the built-in Go
map retrieval operation, which returns 'value, ok'.

The root directory contains a BTree type that wraps the new
"internal" BTree, and provides api compatibilty with the previous
version of this package.