Added Map, Set, and Generic type
The Map uses a key-value pair generic tuple where the key is an
ordered type like string, int, float, etc.

The Set is like Map but with no value. The interface is a little
different too. Has Insert, Contains, and Delete.

The Generic type is the main B-tree type for data that needs a
custom comparator.

The older BTree type is still available for backwards compatibilty.